fix: normal logout affects all sessions (#3571)
Previously all user tokens were deleted, which logouts from all sessions.

+++ framework/core/src/Forum/Controller/LogOutController.php
@@ -102,11 +102,12 @@ class LogOutController implements RequestHandlerInterface
             return new HtmlResponse($view->render());
         }
 
+        $accessToken = $session->get('access_token');
         $response = new RedirectResponse($url);
 
         $this->authenticator->logOut($session);
 
-        $actor->accessTokens()->delete();
+        $actor->accessTokens()->where('token', $accessToken)->delete();
 
         $this->events->dispatch(new LoggedOut($actor, false));
